Prevent clean_up_envs from deleting local development environment

+++ conftest.py
@@ -1,6 +1,7 @@
 import json
 import os
 import posixpath
+import re
 import shutil
 import subprocess
 import sys
@@ -273,7 +274,7 @@ def clean_up_envs():
             conda_info = json.loads(subprocess.check_output(["conda", "info", "--json"], text=True))
             root_prefix = conda_info["root_prefix"]
             for env in conda_info["envs"]:
-                if env != root_prefix:
+                if env != root_prefix and re.search(r"mlflow-\w{16,}$", env):
                     shutil.rmtree(env, ignore_errors=True)
